pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-explorer"
version = "0.2.0"
description = "Agent-agnostic CLI for exploring AI agent chat data (Cursor, Cline, Aider, etc.)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Agent Explorer" }
]
dependencies = [
    "numpy>=2.0.2",
    "toponymy>=0.3.0",
    "umap-learn>=0.5.9.post2",
    # Local development dependencies - use git URLs for public release
    # For local dev, install with: uv pip install -e ../multiscale -e ../llm-helpers
    "multiscale @ git+https://github.com/arclabs561/multiscale.git",
    # llm-helpers is not yet published - users need to install manually or use local path
    # "llm-helpers @ git+https://github.com/user/llm-helpers.git",  # TODO: Publish to GitHub/PyPI
]

[project.optional-dependencies]
llm = [
  "openai>=1.50.0",
]
extras = [
  "python-dotenv>=1.0.1",
]
dotenv = [
  "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
  "pytest>=8.3",
  "pytest-cov>=4.1",
  "pytest-xdist>=3.5",
  "pytest-randomly>=3.15",
  "pytest-sugar>=0.9.7",
  "pytest-timeout>=2.3",
  "ruff>=0.5",
]

[project.scripts]
agent-explorer = "agent_explorer.cli:main"
cursor-explorer = "agent_explorer.cli:main"  # Backward compatibility alias
cursor-chat-explorer = "agent_explorer.cli:main"  # Backward compatibility alias

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["agent_explorer"]
py-modules = ["llm_utils", "llm_cache"]

[tool.uv]
# uv-specific configuration
# Dev dependencies are in [dependency-groups.dev] (PEP 735 standard)

[tool.ruff]
line-length = 220
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["W191", "E501"]
